Move deployment files into new package dirs
